style: apply xo config and rename bad files
