+++ src/exceptions/dom-not-found-error.ts
@@ -1,12 +1,9 @@
 /**
  * DOMNotFoundError exception. For not provided dom resolver.
  */
-export class DOMNotFoundError extends Error {
-    /**
-     *
-     */
-    constructor(type: 'DOMParser' | 'XMLSerializer' | 'DOMImplementation') {
-        super(`No ${type} was provided.`);
-        this.name = `${type}NotFoundError`;
-    }
+export class DomNotFoundError extends Error {
+  constructor(type: 'DOMParser' | 'XMLSerializer' | 'DOMImplementation') {
+    super(`No ${type} was provided.`);
+    this.name = `${type}NotFoundError`;
+  }
 }
