--- src/nodes/c-node-has-value-interface.ts
@@ -1,5 +1,0 @@
-export type CNodeHasValueInterface = {
-    value(): string;
-
-    setValue(value: string): void;
-};
