+++ src/nodes/node-has-value-interface.ts
@@ -1,0 +1,5 @@
+export type NodeHasValueInterface = {
+  value(): string;
+
+  setValue(value: string): void;
+};
