+++ vitest.config.ts
@@ -1,17 +1,16 @@
 import { defineConfig } from 'vitest/config';
+import tsconfigPaths from 'vite-tsconfig-paths';
 
 export default defineConfig({
-    test: {
-        globals: true,
-        alias: {
-            src: './src',
-        },
-        coverage: {
-            all: true,
-            provider: 'istanbul',
-            reporter: ['text', 'lcov'],
-            include: ['src/**/*.ts'],
-        },
-        environmentMatchGlobs: [['**/*.browser.test.ts', 'jsdom']],
+  plugins: [tsconfigPaths()],
+  test: {
+    globals: true,
+    coverage: {
+      all: true,
+      provider: 'istanbul',
+      reporter: ['text', 'lcov'],
+      include: ['src/**/*.ts'],
     },
+    environmentMatchGlobs: [['**/*.browser.test.ts', 'jsdom']],
+  },
 });
